list route

+++ db-api/routes/groups.js
@@ -5,7 +5,7 @@ const db = require('../model/index');
 router.get('/', async (req, res) => {
     try {
         const groups = await db.Group.findAll();
-        res.json(groups);
+        res.status(200).json(groups);
     } catch (err) {
         console.error('Error dans récupération des groupes :', err);
         res.status(500).json({ error: 'Error dans récupération des groupes' });
@@ -18,7 +18,7 @@ router.get('/:groupId', async (req,res)=>{
         if(group === null){
             res.status(404).json({ error: 'Group non trouvé.' });
         }else{
-            res.json(group)
+            res.status(200).json(group)
         }
         
     } catch (error) {
@@ -37,7 +37,7 @@ router.post('/', async (req,res)=>{
             creator_id: creator_id
         });
 
-        res.json(createdGroup);
+        res.status(201).json(createdGroup);
     } catch (error) {
         console.error(`Error dans l'insertion du groupe :`, error);
         res.status(500).json({ error: "Erreur dans l'insertion du groupe." });
@@ -46,10 +46,16 @@ router.post('/', async (req,res)=>{
 
 router.delete('/:groupId', async (req,res)=>{
     try {
+        const existingGroup = await db.Group.findByPk(req.params.groupId)
+
+        if(!existingGroup){
+            return res.status(404).json({error: 'Groupe introuvable'})
+        }
+
         const removeGroup = await db.Group.destroy({
             where: { group_id: req.params.groupId }
         })
-        res.json(removeGroup)
+        res.status(204).json(removeGroup)
     } catch (error) {
         console.error(`Error dans suppression du groupe ${req.params.groupId} :`, error);
         res.status(500).json({ error: 'Error dans suppression du groupe' });
@@ -76,7 +82,7 @@ router.patch('/:groupId', async (req, res)=>{
             where: { group_id: req.params.groupId },
         })
 
-        res.json(patchedGroup)
+        res.status(200).json(patchedGroup)
         
     } catch (error) {
         console.error(`Error dans maj du groupe ${req.params.groupId} :`, error);
